requirements.txt
pandas>=1.5
pyyaml>=6.0
requests>=2.28
pyodbc>=4.0
pytest>=7.0
pandera>=0.18
openpyxl>=3.0
powerbiclient>=0.6.0  # optional: Power BI REST helper
